efficient run wrapper

+++ qsub_pipeline.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+pipeline_root="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
+
 ### Defaults for PBS options ###
 pbs_project="pq84"    # --pbs-project
 pbs_queue="normalbw"  # --pbs-queue
@@ -10,6 +12,11 @@ storage="gdata/u86+scratch/u86+gdata/pq84+scratch/pq84"  # --storage
 mail_opts="abe"       # --mail-opts
 job_name="RNAseq_pipeline" # --job-name
 mail_user="rhys.gillman@jcu.edu.au"          # --mail-user
+do_submit=false
+keep=true
+
+timestamp=$(date +%Y%m%d_%H%M%S)
+pbs_file="qsub_pipeline_${timestamp}.pbs"
 
 ###  Pull out ONLY PBS-flags ###
 run_args=()
@@ -17,15 +24,18 @@ run_args=()
 while [[ $# -gt 0 ]]; do
   case "$1" in
     # PBS flags (consume but don’t forward)
-    --pbs-project)   pbs_project="$2";   shift 2 ;;
-    --pbs-queue)     pbs_queue="$2";     shift 2 ;;
-    --pbs-walltime)  walltime="$2";      shift 2 ;;
-    --pbs-mem)       mem="$2";           shift 2 ;;
-    --pbs-jobfs)     jobfs="$2";         shift 2 ;;
-    --pbs-storage)   storage="$2";       shift 2 ;;
-    --pbs-mail-opts) mail_opts="$2";     shift 2 ;;
-    --pbs-job-name)  job_name="$2";      shift 2 ;;
-    --pbs-mail-user) mail_user="$2";     shift 2 ;;
+    --submit)           do_submit=true;     shift 1 ;;
+    --keep)             keep="$2";          shift 2 ;;
+    --pbs-project)      pbs_project="$2";   shift 2 ;;
+    --pbs-queue)        pbs_queue="$2";     shift 2 ;;
+    --pbs-walltime)     walltime="$2";      shift 2 ;;
+    --pbs-mem)          mem="$2";           shift 2 ;;
+    --pbs-jobfs)        jobfs="$2";         shift 2 ;;
+    --pbs-storage)      storage="$2";       shift 2 ;;
+    --pbs-mail-opts)    mail_opts="$2";     shift 2 ;;
+    --pbs-job-name)     job_name="$2";      shift 2 ;;
+    --pbs-mail-user)    mail_user="$2";     shift 2 ;;
+    --pbs-script-name)  pbs_file="$2";      shift 2 ;;
     # anything else is for the pipeline
     *) run_args+=("$1"); shift ;;
   esac
@@ -49,7 +59,6 @@ fi
 
 ### Generating PBS Submission Script ###
 
-pbs_file=$(mktemp qsub_pipeline.XXXXXXXX.pbs)
 cat > "$pbs_file" <<EOF
 #!/usr/bin/env bash
 #PBS -P $pbs_project
@@ -76,10 +85,22 @@ echo "Job identifier is \$PBS_JOBID"
 echo "Working directory is \$PBS_O_WORKDIR"
 
 cd "\$PBS_O_WORKDIR"
-bash "\$(pwd)/run_pipeline.sh" ${run_args[@]}
+bash "$pipeline_root/run_pipeline.sh" ${run_args[@]}
 EOF
 
-echo "Submitting pipeline job via qsub…"
-jobid=$(qsub "$pbs_file")
-rm -f "$pbs_file"
-echo "Submitted job $jobid"
+
+if [[ "$do_submit" == true ]]; then
+  echo "Submitting pipeline job via qsub…"
+  jobid=$(qsub "$pbs_file")
+  
+  if [[ "$keep" == false ]]; then
+    rm -f "$pbs_file"
+  fi
+  
+  echo "Submitted job $jobid"
+else
+  #mv "$pbs_file" "./$(basename "$pbs_file")"
+  echo "Generated PBS script: $pbs_file"
+  echo "When you’re ready: qsub $pbs_file"
+fi
+
